pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
# Check https://flit.readthedocs.io/en/latest/pyproject_toml.html for all available sections
name = "ansys-mechanical-core"
version = "0.4.dev0"
description = "A python wrapper for ansys mechanical pymechanical"
readme = "README.rst"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "ANSYS, Inc.", email = "pyansys.support@ansys.com"},
]
maintainers = [
    {name = "PyAnsys developers", email = "pyansys.maintainers@ansys.com"},
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "importlib-metadata >=4.0",
    "appdirs>=1.4.0",
    "grpcio>=1.30.0",
    "protobuf>=3.12.2",
    "ansys-platform-instancemanagement>=1.0.1",
]

[tool.flit.module]
name = "ansys.mechanical.core"

[project.urls]
Source = "https://tfs.ansys.com:8443/tfs/ANSYS_Development/Mechanical/_git/pymechanical"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
force_sort_within_sections = true
line_length = 100
default_section = "THIRDPARTY"
src_paths = ["doc", "src", "tests"]

[tool.coverage.run]
source = ["ansys.mechanical"]
# omit = [
#         "src/ansys/mechanical/core/mechanical_pb2.py",
#         "src/ansys/mechanical/core/mechanical_pb2_grpc.py"
#        ]

[tool.coverage.report]
show_missing = true
